* Fix time-should-be-a-floating-point error

+++ include/Hexagon/Shaders/Fun.hpp
@@ -25,7 +25,7 @@ namespace Hexagon
       void set_flat_color(ALLEGRO_COLOR flat_color={});
       void set_texture_width(int width=1);
       void set_texture_height(int height=1);
-      void set_time(int time=1);
+      void set_time(float time=1);
       static std::string obtain_vertex_source();
       static std::string obtain_fragment_source();
       };
